Update README.md
Guess the secret word

Write a program that prompts the user to enter the secret word. Their input should then be compared to a string variable defined in the code and initialized to a secred word. For example:
string secret = "please"
Use an if/else statement to display if the user guessed correctly or not.